CEA_Flask/requirements.txt
bcrypt==4.0.1
blinker==1.5
bunnet==1.1.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-model==0.0.5
MarkupSafe==2.1.2
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
toml==0.10.2
typing_extensions==4.4.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
